check if Pipfile exists

+++ setup_env.sh
@@ -3,5 +3,5 @@
 python3 -m venv venv
 source venv/bin/activate
 pip install pipenv
-pipenv install --dev
+test -f Pipfile && pipenv install --dev
 deactivate
